rename header files so debs are installable (#1443)
