+++ moveit_planners/pilz_industrial_motion_planner/include/joint_limits_copy/joint_limits_rosparam.hpp
@@ -24,7 +24,7 @@
 #include <limits>
 #include <string>
 
-#include "joint_limits/joint_limits.hpp"
+#include "joint_limits_copy/joint_limits.hpp"
 #include "rclcpp/rclcpp.hpp"
 
 namespace
