+++ moveit_planners/pilz_industrial_motion_planner/include/pilz_industrial_motion_planner/joint_limits_extension.h
@@ -36,7 +36,7 @@
 
 #include <limits>
 
-#include <joint_limits/joint_limits.hpp>
+#include <joint_limits_copy/joint_limits.hpp>
 
 #include <map>
 #include <string>
